working unsigned nbrs (lel)

+++ src/string/ft_utoa_base.c
@@ -6,18 +6,18 @@
 /*   By: xperrin <xperrin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/02 22:33:42 by xperrin           #+#    #+#             */
-/*   Updated: 2018/03/05 21:59:29 by xperrin          ###   ########.fr       */
+/*   Updated: 2018/03/06 15:26:10 by xperrin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static	int		ft_cntdigt_b(intmax_t n, size_t radix)
+static	int		ft_cntdigt_b(uintmax_t n, size_t radix)
 {
 	size_t	r;
 
 	r = 1;
-	while (n >= (int)radix)
+	while (n >= radix)
 	{
 		n /= radix;
 		r++;
